#39 refactor: improve to accept passage and phrases
