+++ src/test/java/com/kairos/context_engine/infrastructure/graph/executor/KnowledgeGraphMutationExecutorTest.java
@@ -1,6 +1,5 @@
-package com.kairos.context_engine.infrastructure.graph;
+package com.kairos.context_engine.infrastructure.graph.executor;
 
-import com.kairos.context_engine.infrastructure.graph.KnowledgeGraphMutationExecutor;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
